Fix the crash in Terminal Logger shutdown. (#10163)
Fixes #9662

Context
When we use TL, we save the console output encoding to a variable, change the encoding
to UTF8 for the logging duration and attempt to set it back to the previous value during the Terminal Logger shutdown. It fails in some terminal emulators. It seems to be either a bug of the terminal emulator or Console.OutputEncoding property.

Changes Made
Wrapped the setting of the console output in try-catch block.

Testing
Unit tests