[BuildCheck] Implement double writes analyzer (#10184)
Fixes #9881

### Context

We believe there is value in flagging cases where the same file is written by more than one task during a build. https://github.com/dotnet/source-build/issues/4390 is an example of a recent hard-to-diagnose build issue that could have been prevented by this analyzer.

### Changes Made

This PR adds a new built-in analyzer and makes a few supporting changes.

### Testing

- Existing and new unit tests.
- Built a few larger repos with the analyzer enabled.

### Notes

The changes contain a fix/workaround for #10176.