Permit erroring on missing project reference Fixes #7528 (#10367)
Fixes #7528

Context
If a project reference does not exist, we currently throw a warning. It was requested in #7528 that there should at least be an option to make it an error.

Changes Made
Throws a warning or an error, depending on the ErrorOnMissingProjectReference property, if a ProjectReference does not exist.

Testing
Tried building a project with a nonexistent project reference with and without that property set to True.