Fix wrong paths in SharedOutputPathAnalyzer. (#10472)
Related to #10414

Context
PR #10238 added normalization of the path (which includes getting full path) before the project directory path is added to it. This led to the relative path being turned into a full path using CurrentDirectory instead of the project directory. This leads to enormous amount of false-positive analyzer's warnings (400 instead of 3) that cause the hang. This PR fixes the bug with paths and does not fix actual cause of a hang.

Changes Made
Reverted the PR #10238 and instead added normalization of the path later in the code when the project path is already combined with the relative path.

Testing
unit tests and local runs