Sync Microsoft.Build version in analyzer template with Version.props in the workflow (#10345)
Fixes #9923

Context
#9789 added a template for custom analyzer. Microsoft.Build version specified in the template should be synced with Version.props during packing. See https://github.com/dotnet/msbuild/pull/9789/files#r1521218723 for more details.
This needs to be done after #9915.

Changes Made
Create a workflow. When there are updates in the file Version.props, the flow will be trigger and check the version. If the version in the template is different form the VersionPrefix in the Version.props. It will create a PR to sync the version between them.