Add Build Submission Started event (#10424)
Fixes #10145

Context
It is currently not possible to check for global variables during build execution. To have access to that data it needs to be passed on events to the processing nodes. None of the events we had previously were good matches to pass along this information, so it was decided to create one for build submission started, as this point in the build we have all properties loaded with the correct value, but its not too late to make use of them.

Changes Made
Added a new BuildSubmissionStartedEventArgs based on BuildStatusEventArgs and added it to the event handlers.
A copy of the enum BuildRequestDataFlags was added to Microsoft.Build.Framework.