Suppress CodeQL for BinaryFormatter in .resx (#10570)
The default .NET Framework build process deserializes resources in
MSBuild, then reserializes them in an expected format for use at
runtime. This is still supported, even though the dominant serialization
mechanism in these projects is `BinaryFormatter`, to maintain build
compatibility with older projects. The need to trust resources embedded
in your application is documented in the MSBuild documentation.