BuildCheck has better message when Check fails on Initialize (#10612)
Partial fix of #10522

Context
When a Check throws an exception it falls with an internal logger exception and crashes the build. This is not ideal. It decided that when a Check fails, we do not fail the build, just give a warning and deregister the check.

Changes Made
Changed the BuildCheck logger to catch initialization exceptions, and dispatch them as warnings and not errors.

Testing
Added end to end test with a sample custom check that throws an exception.

Notes
This is just a catch for exception on Check initialization. There are tests added for other cases but the fix will be different for them, so it will be done in another PR.