Removed scheduler debugger from MSBuildDebugEngine (#10868)
Fixes #10639

Context
There was an internal ticket that seems to originate from the scheduler logger. The team uses the engine debugger variable to keep track of their builds. However, it seems that the scheduler logger does not always work as expected and it is turned on when using MSBuildDebugEngine. Since we think there isn't a lot of information that is necessary from scheduler logger, we are trying to disable it when using the MSBuildDebugEngine variable, and leaving it on when explicitly turned on.

Changes Made
Does not turn on scheduler debugger when using MSBuildDebugEngine