Error handling for failures on registered actions for Custom Checks (#10914)
Fixes #10522

Context
Second PR for that issue. This PR covers the cases in a Custom Check where there is an exception while registering actions, and when running said actions.
Since actions are individuals, we do not deregister the whole analyzer when one fails.

It has been decided to expose these errors are warnings instead to not break the build where they show up.

Changes Made
Added some try catch statements so errors do not break builds. Messages have been added to each case for more detail as we do not preserve anything else from the exception.