ToolTask UseUtf8Encoding value of 'true' is equal to 'always' (#10949)
Fixes #10907

### Context
The `UseUtf8Encoding` property of the `Exec` task and other `ToolTask` derived tasks should accept `true` and equivalent to `always`.

### Changes Made
`ToolTask` doesn't validate or interpret the `UseUtf8Encoding` property. The value of the property is passed to and interpreted by `EncodingUtilities.BatchFileEncoding()`.

`EncodingUtilities` was modified to add a const string for `TRUE` and the `BatchFileEncoding()` method was modified to handle `TRUE` and `ALWAYS` the same.

### Testing
Created a unit test that compares the returns for `ALWAYS` and `TRUE`, confirmed that the unit test failed, modified the code, and confirmed the unit test passes.

Tested on Windows.

### Notes
On non-Windows platforms the `UseUtf8Encoding` property has no effect.