Build against old versions of maintenance-packages
Avoids problems caused by updating references in MSBuild for core types
that are likely to be used in applications that use MSBuildLocator to
call our API: if we update our baked-in references, we will break some
of them, including Roslyn's MSBuildWorkspace BuildHost. If instead we
bake in the same references we've been using, but ship and binding
redirect to the newer versions, we can update (at runtime in normal
operation) while still being functional in older executables that
ship a stale version (and thus don't binding redirect to a new version).