Merged PR 627368: [vs16.11] DownloadFile should not rely on the response headers
DownloadFile should not rely on the remote server response headers. Unless the DestinationFileName task parameter is specified - let's just fallback to the request URI - which is as well the publicly documented behavior

----
#### AI description  (iteration 2)
#### PR Classification
Bug fix

#### PR Summary
This pull request updates the `DownloadFile` task to avoid relying on response headers for determining the file name.
- `src/Tasks/DownloadFile.cs`: Modified `TryGetFileName` method to use the request URI instead of the response headers to determine the file name. Updated method calls accordingly.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->