Tolerate duplication in p2p negotiation batches
The intent of batching over the condition
    '@(ProjectsWithNearestPlatform)' == '%(Identity)'
is to run the update for each item in the list (that meets the rest of
the condition).

However, if there are two list entries with identical `%(Identity)`--in
this case duplicate ProjectReferences--they batch into the same bucket
and then that condition is `false`.

Instead, batch over unique identities to tolerate duplicates.

Fixes #2688.

Co-Authored-By: Benjamin Witman <160950603+benwitmanmsft@users.noreply.github.com>