Support testhost.exe in TaskHost-launching tests (#11956)
Some tests were failing when run from Visual Studio because they were
running in a .NET 9 process launched as `TestHost.exe`, not in a `dotnet.exe`
host. That caused TaskHost launching to try to run `TestHost.exe MSBuild.dll`
which failed, as TestHost isn't a general-purpose launcher.

Add another fallback to `GetCurrentHost` to fall back to the current
shared framework's parent `dotnet.exe`, if we can find it.