Perf: Implicit Lookup.Scope stacks (#12060)
Reduces `LinkedList` related allocations in `Lookup`.

`Lookup` uses `LinkedList<Scope>` to implement a stack with the ability to peek into the next item. Given how many of these are created, this is a large source of allocations since every element requires an overhead of 48 bytes per `LinkedListNode<T>`.

This isn't necessary since we already have an object which represents a unidirectional node... the `Scope` itself.
 
This adds a `Parent` property to `Lookup.Scope` which allows `Lookup` to use it as an implicit stack, avoiding the need for any intermediate collections.