Make TerminalLogger resilient to odd Console configurations, and adjust rendering for some verbose data (#12083)
Fixes #12081, #11213, #10979.

### Context

The Terminal Logger could be enabled in cases where automated systems
mimicked a tty, but did so with operating parameters that caused us to
crash at runtime. While I was in here I also made a few small changes to
the way some specific messages were reported (or not reported!)

### Changes Made

This PR
* strengthens the TL auto-enablement checks to make it less likely that
TL is enabled in those environments
* makes the Terminal wrapper more resilient to the zero-width and
zero-height scenarios in case users force TL on in such environments
* Slightly adjusts the formatting of the SDK preview message to render
_more like_ a normal diagnostic, but without some of the file-based
cruft of that form. It also enforces a once-per-build reporting of this
message.
* Prevents XMake.cs from reporting the raw MSBuild commandline on Stdout
when TL is in use.

### Testing

Manual testing only, would love some ideas - maybe synthetic Terminal
instances?

### Notes