Perf: Concurrent ConfigCache (#12052)
Context
Right now, ConfigCache uses a system of two backing Dictionary instances protected by a lock to ensure that any operations atomically update both collections, and that concurrent reads don't fail:

private IDictionary<int, BuildRequestConfiguration> _configurations;
private IDictionary<ConfigurationMetadata, int> _configurationIdsByMetadata;
private object _lockObject = new object();

Given that the most frequent operations here are reads, this causes a significant amount of lock contention as seen above.

In some cases the dictionaries are replaced together or modified in a way that can't be gated via a simple TryX method - so to replace with ConcurrentDictionary, we need some additional steps to ensure that an unexpected race condition doesn't cause the cache to go out of sync.

Changes Made
Methods which perform multiple operations on the dictionaries take a local reference so that they're guaranteed to be operating on the same instance (in the event that it gets swapped out mid-execution).
Both dictionaries are wrapped in a instance container so that methods which do a replace, clear, or serialize can do so atomically.
private record class Configurations : ITranslatable
{
    private ConcurrentDictionary<int, BuildRequestConfiguration> _byId;
    private ConcurrentDictionary<ConfigurationMetadata, BuildRequestConfiguration> _byMetadata;
}
Double adds / removes explicitly throw if an unrecoverable race is detected. Note that the current AddConfiguration() implementation already throws if an existing ID is added, so this isn't expected behavior.
e.g. ById.TryAdd() == true -> ById.TryRemove() == true -> ByMetadata.TryRemove() == false -> ByMetadata.TryAdd() == true

Additional Changes
Enumerations now go through the full KeyValuePair enumerator, as both Keys and Values on ConcurrentDictionary will create a new array and copy a snapshot, which is different than Dictionary behavior. This enumerator is threadsafe.
Translate just serializes the ID dictionary. Both can be derived based on BuildRequestConfiguration, so it doesn't really matter which one is picked.

Co-authored-by: Tomas Bartonek <tom.bartonek@volny.cz>