Don't log false-condition expansion unnecessarily
This came up in investigating devdiv2535034, which was a regression in
exception count caused by a property function within a condition throwing
an exception when fully stringified (but not on normal expansion due to
short-circuiting condition evaluation).

But the problem is more broad--during a design-time build or even a regular
build at low verbosity, these low-pri events would just be thrown away.

Extend the OnlyLogCriticalEvents check to include verbosity for the
logging-only expansion of a false condition.