Switch ItemBucket to be a struct (#12148)
Fixes #

### Context

The `ItemBucket` type is currently a class that gets allocated fairly
frequently, added to collections, and passed around. Otherwise, the
class is fairly a straightforward container of a handful of items.
Switching it to a struct would avoid the heap allocations here.

Before:

<img width="880" height="81" alt="image"
src="https://github.com/user-attachments/assets/fdba1c32-3305-401f-b97a-bc7f7cf0db70"
/>

After:

<img width="877" height="59" alt="image"
src="https://github.com/user-attachments/assets/4dea6232-976a-48a1-a3cb-ff73163db12d"
/>

Some things to note about the changes:

I needed to switch to having `ItemBucket` implement the generic
`IComparable<T>` to avoid boxing the struct.

A class is the size of a pointer and structs can vary in size depending
on how many fields are in the struct. This is why there is an *increase*
in the `ItemBucket[]` allocations that back the `List<ItemBucket>`.

To reduce the number of `ItemBucket[]` allocations, I reduced the number
of times we need to resize by utilizing `EnsureCapacity()`. While newer
.NET code has this, I had to create an extension function for Framework.

There was an additional `List<ItemBucket>` that was being allocated just
to sort the buckets in `BucketSequenceNumber`. We can do this in place
to avoid the allocation.


### Changes Made


### Testing


### Notes