Perf: Pool scheduler request buffers (#12007)
### Fixes

Heavy allocations due to scheduler temporary buffers.

### Context

`Scheduler` allocates a ton of temporary lists / arrays to dump the
current list of requests. This is necessary because of the
semi-recursive nature of the call stack. As you see here, the largest
source of allocations `ResumeRequiredWork` ~can call back into
`ResumeRequiredWork`~, which would a runtime error without the buffer as
the call stack may modify the request list.

Edit: flubbed that part a bit - it's not methods calling into themselves
but multiple methods which take an enumeration calling other methods
which modify the target list.


![image](https://github.com/user-attachments/assets/c9948e59-dc49-4572-8afc-1f7d94790fa9)

### Changes Made

This creates an instance-level `ArrayPool` to share buffers across
methods.

Specifically this avoids `ArrayPool.Shared` since the implementation
differences between `SharedArrayPool` and `ConfigurableArrayPool` cause
the shared pool to evict the pooled arrays (my guess is that we end up
exceeding the threshold due to recursive calls).

The `ArrayPool` is thrown out and recreated after each build via
`BuildManager.Reset()`, so we're not just holding on to a bunch of extra
memory.

After this, we end up with:


![image](https://github.com/user-attachments/assets/38428e7a-6749-496d-a008-da08dea122df)

The runtime cost of the pool is basically trivial:


![image](https://github.com/user-attachments/assets/93c00b3d-192a-4e54-b7d8-3edbe7287132)

---------

Co-authored-by: Tomas Bartonek <tom.bartonek@volny.cz>