Make TerminalLogger capable of acting as a ForwardingLogger, and use it to implement SourceRoot-based relative paths for out-of-tree projects (#12082)
Fixes https://github.com/dotnet/msbuild/issues/9806

### Context

This is another take on making TL a forwarding logger. I've gotten it
much better this time - it's working in multi-node setups consistently
on my machine, and I was able to get the Target Outputs for targets that
I care about after a bit of work. We can make this more efficient be
filtering down the Target Outputs _at the forwarding level maybe_? Yep!
That was easy enough to do.

### Changes Made

* Made a Forwarding Terminal Logger
* Hooked it up in the distributed logger registration for multi node
scenarios
* Plumbed through a knob for enabling Target Outputs to be logged in the
build parameters, node configuration, logging server, etc. and toggled
that knob on for TL-using scenarios
* Used this infra to accurately get the Source Roots for the projects,
so that when builds happen outside of the CWD and inside the Source Root
we can render a relative path between the CWD and the Source Root as the
'anchor' for the project outputs
* From this base we can do many of the data-based enhancements (NuGet
package detection?!) that we've wanted to do.

### Testing

Manual testing - the test case is building a solution from a project
that is a _sibling_ of that solution (so same parent, but different
directory), building with `/m:8`.


![image](https://github.com/user-attachments/assets/65436318-2809-4428-89bc-0937f9d1225e)

### Notes

TODO:
* Tests
* Perf comparison? This will likely be a mixed bag because we _should_
be emitting fewer events from the child loggers, but some of the events
may contain more data.
* Reviews/cleanup

---------

Co-authored-by: YuliiaKovalova <95473390+YuliiaKovalova@users.noreply.github.com>
Co-authored-by: YuliiaKovalova <ykovalova@microsoft.com>