Add overload for ProjectCollection constructor to prevent binary breaking change (#12296)
### Context
https://github.com/dotnet/msbuild/pull/12082 introduced a
binary-breaking change to the ProjectCollection constructors, which is
breaking VMR codeflow at https://github.com/dotnet/dotnet/pull/1764.

### Changes Made

Made the new change in a binary-compatible way by adding the old
constructor signature and forwarding to the new constructor with a
default value set for the new parameter. This also meant that we get to
remove the CompatibilitySuppression, which should have been a red flag
in the first place.