Avoid boxing and unnecessary array allocations (#12175)
There are several delegates called during build that cause some sneaky
allocations. For instance, there is boxing of the `BuildRequestEngineStatus` enum.

The culprit is the `ErrorUtilities.VerifyThrow()` calls that take
`object` as parameters for the formatted message. To make the call, the
enum gets boxed even when the condition check returns `true`. Adding
additional overloads that take the specific type lets multiple callers
benefit, and I did some additional cleanup in the delegates to remove
some allocations.