issue-12415 increase stack size (#12426)
Fixes #12415

### Context
During one of our recent optimization runs we reduced stack size of the
DrainPacketQueue times.
While this is fine in a normal scenario, it sometimes results in a
StackOverflow during debugging or performance runs.

### Changes Made
Increased the stack size.

### Testing


### Notes
@Erarndt is planning to refactor this away in the near future so I hope
this is a sufficient hotfix for the interim.
This doesn't affect production, only custom builds.