Add direct lookup for task assembly (#12516)
Fixes https://github.com/dotnet/msbuild/issues/12514

### Context
NET Task host executes tasks in a separate process.
The parent one loads assembly with the requested task by default and
traverses all the types in it.
Since the requested assembly belongs to net-runtime, some of the
dependencies aren't available as a part of msbuild.exe (parent process)

### Changes Made
1. Add a short path to search for the requested type in the assembly.
2. wrap the dependencies traversal with try/catch. If any of them isn't
available in the parent process - it's expected, since it will be
resolved in dotnet.exe.