Allow NodeProviderOutOfProcTaskHost to manage multiple nodes instead of one per arch (#12577)
Fixes #12552

### Context
Currently, for multithreaded node we cannot force all tasks to run in
task hosts by setting `MSBUILDFORCEALLTASKSOUTOFPROC=1`. It crashes. The
problem is that NodeProviderOutOfProcTaskHost is one per process, and
manages up to 4 nodes for specific architecture.

### Changes Made
1. Propagated `nodeId` if the in-proc nodes (only when multi-threaded
mode is on) into `NodeProviderOutOfProcTaskHost`
2. Used `nodeId` with the combination of `(int)HanshakeOptions` (is
originally used as key for `_nodeContexts`), to generate a unique id for
task host node


### Testing
Manually tested by building this project that references other projects
with `Message` tasks. Modified `Message` task to show the process id it
is running on.
`msbuild.exe .\MainProject.proj - m` with combination of with/without
`-mt` switch and with/without setting `MSBUILDFORCEALLTASKSOUTOFPROC`.

### Notes

---------