allow a parameter to disable the live-updating nodes display (#12581)
Fixes part of #12486, but not all.

### Context

This introduces a new Terminal Logger Parameter (`DISABLENODEDISPLAY`)
that disables the live-updating node display portion of the Terminal
Logger output. This output is very wasteful when LLMs are running the
build, and the hope is that we can toggle this flag automatically at the
dotnet CLI level to maintain the structure benefits of TL while also not
flooding the context.

### Changes Made

* introduced a new parameter and threaded it through the logic

### Testing

* Manual testing done - verified that the flag does in fact disable the
display entirely.


https://github.com/user-attachments/assets/f6e9f74b-77c2-46be-a1ae-7d974d5d7348

### Notes

This will use the logic from https://github.com/dotnet/sdk/pull/51054 in
the SDK to turn off the live-update feature.

Co-authored-by: YuliiaKovalova <95473390+YuliiaKovalova@users.noreply.github.com>