Stop updating VS PackageReferences on insertion
VSSDK packages have recently exposed a "live" dependency on MSBuild,
which has caused several complications in the VS/MSBuild/VSSDK release
process. In consultation with the VSSDK team, the plan for now is to pin
to MSBuild's 18.0 package, so there's a known stable version to use.

VSSDK consumers will have the option to locally reference a higher
MSBuild but we suspect they are unlikely to need to do so.

This can't be a full and final solution because eventually we'll
introduce new API for VS to use and need to reference it in VS. At that
time we'll need to disentangle the "public" and "internal" MSBuild
references in VS.