Fix escaping for dotnet paths with spaces when spawning out-of-proc nodes (#12644)
### Context
When a path to dotnet.exe or MSBuild.dll contains spaces, it's
incorrectly escaped and causes errors like this:
Could not execute because the specified command or file was not found.
* You intended to execute a .NET program, but dotnet-C:\Program does not
exist.
The path gets truncated at the space character, causing the process
launch to fail.

The important notice – it’s a critical scenario for the feature
dogfooding , since dotnet.exe usually is placed in “Program Files” (path
with spaces).

It wasn’t caught during the development because the change required
coordination with SDK and the default private SDK was in a path
without spaces, unlike `C:\Program Files`.

### Changes Made
Added proper escaping for the msbuild.dll path to ensure it works
correctly even when the path contains spaces.

### Regression
No, it’s a bug in the new feature

### Risks
No risks, minimal well-boxed change

### Testing
UT + manual

### Notes
This logic is taken from https://github.com/dotnet/msbuild/pull/12630