Add missing IsUnitTestProject to StringTools.UnitTests.net35 (#12697)
Arcade sets this property by naming convention, which doesn't apply to
this project.

See
https://github.com/dotnet/arcade/blob/f11f9fc40850209eb01a890dabd5e7d7506c7b4e/src/Microsoft.DotNet.Arcade.Sdk/tools/Tests.props#L16

This project was previously not run in CI at all because of this.

This is from the latest build on main:

<img width="1898" height="1080" alt="image"
src="https://github.com/user-attachments/assets/2cab591b-e91d-4a51-9373-a57d961f61eb"
/>

which makes sense based on the condition in
https://github.com/dotnet/arcade/blob/f11f9fc40850209eb01a890dabd5e7d7506c7b4e/src/Microsoft.DotNet.Arcade.Sdk/tools/Tests.targets#L35

Additionally, this can easily be verified by looking at build artifact
and noticing that there are no trx files for this project.

This also has an impact on signing the assembly:


https://github.com/dotnet/msbuild/blob/60fc61c383f640985fc62f05bf9cfa01c4159a6b/src/Directory.Build.targets#L7-L16

(as well as collecting code coverage)

In addition, any property conditions around `IsTestProject` needs to be
during targets. It's very early to do that in props, because the
property can be altered by the project itself.