CmdLine parsing was extracted from XMake and the implementation is visible to dotnet (#12704)
### Context
In some cases dotnet needs to process MSBuild cmdline switches including
response files.

### Changes Made
The parsing logic was pulled out from XMake to isolate it in the
internal type, making it easily reusable in the .NET SDK project.

### Testing
All existing tests are passing.

### Notes
The goal of this PR isn’t to refactor the entire API—that’s something we
can consider tackling later on.