RAR node: Buffer log events to client (#12558)
Implements logging for out-of-proc RAR node.

### Context

Since RAR is often the largest source of events in MSBuild, some
complexity is needed to avoid introducing
overhead from excessive allocations and task context-switching. The main
goals here are:

- Logging an event should not block RAR, as IPC can be performed async.
- Events should be batched to reduce the number of IPC calls. However,
too large of a batch will add latency to
the final response packet.
- Pending events should be flushed when RAR completes, but before
sending the result packet.
- Buffers should be resusable between packets, as batches are a fixed
size.
- The processing task should only wake when a full packet is ready, so
that each read of the event channel
will immediately return a struct ValueTask.

As such, two channels are used: one for events, and one for the batch
size.

When an event is queued, a pending count is incremented. When the count
reaches the batch threshold, or the queue
is explicitly flushed, the count is written to the count channel and
reset.

Once RAR completes, events are flushed with the remaining count.

---------

Co-authored-by: YuliiaKovalova <95473390+YuliiaKovalova@users.noreply.github.com>