Add test summary always in terminal logger (#12801)
Fixes https://github.com/microsoft/vstest/issues/15307

### Context
The terminal node can be null before we receive the finish message. This
will cause that some summaries are not accounted for and will be missing
in the final output.

### Changes Made
Always update data for messages that don't require active node.

### Testing
Tested manually on local build of msbuild.

I was not able to trace why the node becomes null, please help.

### Notes