Patch Metadata.json with optprof source
The AzDO task `MicroBuildSigningPlugin` internally uploads the
`Metadata.json` file that it creates, but since we do drop discovery
the Arcade way later, that file didn't have the LKG info. But the
use-LKG code in `.opt-prof.yml` found it explicitly, not by convention,
so patch a _copy_ of the file and then use that.