[vs17.14] Add test summary always in terminal logger (#12852)
**Description**

Running tests with dotnet test, users see inconsistent number of
projects and tests reported in the view. This is a UI only issue,
failing tests will still fail the build. It is still worth fixing the
issue because it makes the runs show repeatable numbers of tests to the
user.

The terminal node can be null before we receive the finish message. This
will cause that some summaries are not accounted for and will be missing
in the final output.

The node is null for this reason:
https://github.com/dotnet/msbuild/issues/12819 That fix is not
backported.

**Customer Impact:** Multiple runs of the same test suite report
different total counts of tests and projects.

**Regression?** No, this was never working correctly since the beginning
of test reporting in Terminal Logger.

**Risk:** Low, it impacts UI only.

Link the PR to the original issue and to the PR to main. 
PR to main is merged: https://github.com/dotnet/msbuild/pull/12801
Fixes https://github.com/microsoft/vstest/issues/15307

No additional impact on packages or localization.